+++ src/BlockLayeredProductSearchProvider.php
@@ -1,5 +1,7 @@
 <?php
 
+require_once __DIR__ . DIRECTORY_SEPARATOR . 'BlockLayeredFiltersConverter.php';
+
 use PrestaShop\PrestaShop\Core\Business\Product\Search\ProductSearchProviderInterface;
 use PrestaShop\PrestaShop\Core\Business\Product\Search\ProductSearchContext;
 use PrestaShop\PrestaShop\Core\Business\Product\Search\ProductSearchQuery;
@@ -13,121 +15,12 @@ use PrestaShop\PrestaShop\Core\Business\Product\Search\FacetsURLSerializer;
 class BlockLayeredProductSearchProvider implements ProductSearchProviderInterface
 {
     private $module;
+    private $filtersConverter;
 
     public function __construct(BlockLayered $module)
     {
         $this->module = $module;
-    }
-
-    public function getFacetsFromBlockLayeredFilters(array $blockLayeredFilters)
-    {
-        $facets = [];
-        foreach ($blockLayeredFilters as $facetArray) {
-            $facet = new Facet;
-            $facet
-                ->setLabel($facetArray['name'])
-                ->setMultipleSelectionAllowed(true)
-            ;
-            switch ($facetArray['type']) {
-                case 'category':
-                case 'quantity':
-                case 'condition':
-                case 'manufacturer':
-                case 'id_attribute_group':
-                case 'id_feature':
-                    $type = $facetArray['type'];
-                    if ($facetArray['type'] == 'quantity') {
-                        $type = 'availability';
-                    } elseif ($facetArray['type'] == 'id_attribute_group') {
-                        $type = 'attribute_group';
-                        $facet->setProperty('id_attribute_group', $facetArray['id_key']);
-                    } elseif ($facetArray['type'] == 'id_feature') {
-                        $type = 'feature';
-                        $facet->setProperty('id_feature', $facetArray['id_key']);
-                    }
-                    $facet->setType($type);
-                    foreach ($facetArray['values'] as $id => $filterArray) {
-                        $filter = new Filter;
-                        $filter
-                            ->setType($type)
-                            ->setLabel($filterArray['name'])
-                            ->setMagnitude($filterArray['nbr'])
-                            ->setValue($id)
-                        ;
-                        $facet->addFilter($filter);
-                    }
-                    break;
-                case 'weight':
-                case 'price':
-                    $facet->setType($facetArray['type']);
-                    $facet->setProperty('min', $facetArray['min']);
-                    $facet->setProperty('max', $facetArray['max']);
-                    $filter = new Filter;
-                    $filter
-                        ->setType($facetArray['type'])
-                        ->setValue([
-                            'from' => $facetArray['values'][0],
-                            'to' => $facetArray['values'][1],
-                        ])
-                    ;
-                    break;
-            }
-            $facets[] = $facet;
-        }
-        return $facets;
-    }
-
-    public function getBlockLayeredFiltersFromFacets(array $facets)
-    {
-        $blockLayeredFilters = [];
-
-        foreach ($facets as $facet) {
-            switch ($facet->getType()) {
-                case 'category':
-                case 'availability':
-                case 'condition':
-                case 'manufacturer':
-                case 'attribute_group':
-                case 'feature':
-                    $type = $facet->getType();
-                    if ($type === 'availability') {
-                        $type = 'quantity';
-                    } elseif ($type === 'attribute_group') {
-                        $type = 'id_attribute_group';
-                    } elseif ($type === 'feature') {
-                        $type = 'id_feature';
-                    }
-                    $blockLayeredFilters[$type] = [];
-                    foreach ($facet->getFilters() as $filter) {
-                        if (!$filter->isActive()) {
-                            continue;
-                        }
-                        $key    = count($blockLayeredFilters[$type]);
-                        $value  = $filter->getValue();
-                        if ($type === 'id_attribute_group') {
-                            $key = $value;
-                            $value = $facet->getProperty('id_attribute_group').'_'.$filter->getValue();
-                        }
-                        if ($type === 'id_feature') {
-                            $key = $value;
-                            $value = $facet->getProperty('id_feature').'_'.$filter->getValue();
-                        }
-                        $blockLayeredFilters[$type][$key] = $value;
-                    }
-                    break;
-                case 'weight':
-                case 'price':
-                    $filters = $facet->getFilters();
-                    if (!empty($filters)) {
-                        $blockLayeredFilters[$facet->getType()] = [
-                            $filters[0]->getValue()['from'],
-                            $filters[0]->getValue()['to']
-                        ];
-                    }
-                    break;
-            }
-        }
-        return $blockLayeredFilters;
+        $this->filtersConverter = new BlockLayeredFiltersConverter;
     }
 
     public function addFacetsToQuery(
@@ -140,7 +33,7 @@ class BlockLayeredProductSearchProvider implements ProductSearchProviderInterfac
         $facetAndFiltersLabels = $urlSerializer->unserialize($encodedFacets);
 
         $filterBlock    = $this->module->getFilterBlock();
-        $queryTemplate  = $this->getFacetsFromBlockLayeredFilters(
+        $queryTemplate  = $this->filtersConverter->getFacetsFromBlockLayeredFilters(
             $filterBlock['filters']
         );
 
@@ -171,7 +64,7 @@ class BlockLayeredProductSearchProvider implements ProductSearchProviderInterfac
         $order_by     = $query->getSortOrder()->toLegacyOrderBy(true);
         $order_way    = $query->getSortOrder()->toLegacyOrderWay();
 
-        $blockLayeredFilters = $this->getBlockLayeredFiltersFromFacets(
+        $blockLayeredFilters = $this->filtersConverter->getBlockLayeredFiltersFromFacets(
             $query->getFacets()
         );
 
@@ -196,7 +89,7 @@ class BlockLayeredProductSearchProvider implements ProductSearchProviderInterfac
         $result->setPaginationResult($pagination);
 
         $filterBlock = $this->module->getFilterBlock();
-        $facets      = $this->getFacetsFromBlockLayeredFilters(
+        $facets      = $this->filtersConverter->getFacetsFromBlockLayeredFilters(
             $filterBlock['filters']
         );
